remove the repeated code
