+++ lib/glific/flows/translate/translate.ex
@@ -31,4 +31,24 @@ defmodule Glific.Flows.Translate.Translate do
     {:ok, result} = translate([orig], src, dst)
     hd(result)
   end
+
+  @doc """
+  Cleanup up string list replacing long text exceeding token threshold with warning
+  This reverses the order of string which is reversed again in next function
+  """
+  @token_chunk_size 200
+  @spec check_large_strings([String.t()]) :: [String.t()]
+  def check_large_strings(strings) do
+    strings
+    |> Enum.reduce([], fn string, acc ->
+      # we ca use the gptTokenizer to count the token
+      string_size = Gpt3Tokenizer.token_count(string)
+
+      if string_size > @token_chunk_size do
+        ["translation not available for long messages" | acc]
+      else
+        [string | acc]
+      end
+    end)
+  end
 end
